+++ src/components/Card/index.tsx
@@ -3,16 +3,16 @@ import Style from './style';
 
 type CardProps = {
   active: boolean;
-  index: number;
-  handleCardMouseDown: (index: number) => void;
+  id: string;
+  handleCardMouseDown: (id: string) => void;
   title: string;
 };
 
-const Card = ({ active, index, handleCardMouseDown, title }: Readonly<CardProps>): JSX.Element =>
+const Card = ({ active, id, handleCardMouseDown, title }: Readonly<CardProps>): JSX.Element =>
   <Style
     className={`card${active ? ' active' : ''}`}
     onClick={(event: React.MouseEvent<HTMLInputElement>): void => event.stopPropagation()}
-    onMouseDown={(): void => handleCardMouseDown(index)}>
+    onMouseDown={(): void => handleCardMouseDown(id)}>
     {title}
   </Style>;