+++ src/components/PlayArea/spec.tsx
@@ -9,7 +9,7 @@ import PlayArea from '@components/PlayArea';
 
 describe('PlayArea', () => {
   const setActiveCard = jest.fn();
-  const PlayAreaMock = <PlayArea activeCard={-1} setActiveCard={setActiveCard} />;
+  const PlayAreaMock = <PlayArea activeCard={''} setActiveCard={setActiveCard} />;
 
   it('should render correctly', () => {
     const tree = renderer
@@ -21,7 +21,7 @@ describe('PlayArea', () => {
   describe('Cards', () => {
     it('should render the correct amount of Card components', () => {
       const wrapper = mount(PlayAreaMock);
-      expect(wrapper.find(Card)).toHaveLength(4);
+      expect(wrapper.find(Card)).toHaveLength(6);
     });
   });